Addition of Basic Start Menu Scene
Currently, Start Button loads scene 1, assuming the menu branch is build index 0 and our game starts at build index 1.  This should be changed once we have a definitive scene for the beginning of the game.  Text is basic and background is currently the default, though can be updated at a later time with a different background, we could maybe put one of the levels in the background of the menu.